requirements.txt
Django==1.11.7
django-fe-core==0.2.0
pytz==2017.3
